Add `flea_fee` field to `PriceData`
See https://github.com/the-hideout/tarkov-api/blob/main/resolvers/itemResolver.js#L134
for how the tarkov.dev API calculates the fee. They assume that the
hideout intelligence center is not level 3 and that the hideout
management skill is 0.

This field is just added for the sake of completeness. The idea is to
let RatScanner calculate the fee using the same logic as the link above.

*/